requirements.txt
Flask==3.0.3
pymongo==4.6.1
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.2.0
dnspython==2.6.1
gunicorn==21.2.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5